% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound.R
\name{aba_to_mi_lower}
\alias{aba_to_mi_lower}
\title{Lower bound for ABA as a function of information}
\usage{
aba_to_mi_lower(k, aba)
}
\arguments{
\item{k}{Number of classes}

\item{aba}{average Bayes accuracy}
}
\description{
Lower bound for ABA as a function of information
}

