% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{assume_iid}
\alias{assume_iid}
\title{Produces an identity matrix}
\usage{
assume_iid(X, ...)
}
\arguments{
\item{X}{design matrix}
}
\value{
Estimate for covariance matrix
}
\description{
A \code{Sigma_t_method}.
}

