% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{assume_iid_B}
\alias{assume_iid_B}
\title{Produces an identity matrix}
\usage{
assume_iid_B(Y, sigma2 = 1, ...)
}
\arguments{
\item{Y}{response matrix}

\item{sigma2}{Scaling of matrix}
}
\value{
Estimate for Sigma_b, a diagonal matrix
}
\description{
A \code{Sigma_b_method}.
}

