% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_basis.R
\name{bdwid_all_preds}
\alias{bdwid_all_preds}
\title{Get predictions for a list of basis sets}
\usage{
bdwid_all_preds(accs_sub, basis_sets, nonnegative = TRUE)
}
\arguments{
\item{accs_sub}{subsampled accuracies ATA_k}

\item{basis_sets}{a list of bases obtained e.g. using lapply and get_gaussian_basis_mat}

\item{nonnegative}{constrain coefficients of basis to be nonnegative? default TRUE.}
}
\description{
Get predictions for a list of basis sets
}
