% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassExReg.R
\name{ClassExReg1nn}
\alias{ClassExReg1nn}
\title{Classification Extrapolation with Regression (and bootstrap model selection) for 1 nearest neighbor}
\usage{
ClassExReg1nn(Xm, Ym, i_chosen, basis_sets, Ktrain = 2:nrow(Xm), nboot = 25,
  nonnegative = TRUE)
}
\arguments{
\item{Xm}{Matrix of centroids, rows are centroids and columns are dimensions.}

\item{Ym}{Data matrix, rows are observations and columns are dimensions.}

\item{i_chosen}{True labels.}

\item{basis_sets}{a list of bases obtained e.g. using lapply and get_gaussian_basis_mat.
If there is only one element in the list, then model selection is not applied.}

\item{Ktrain}{the number of classes k for which we have computed average test accuracies ATA_k.}

\item{nboot}{Number of bootstrap samples to use in model selection.}

\item{nonnegative}{constrain coefficients of basis to be nonnegative? default TRUE.}
}
\description{
Classification Extrapolation with Regression (and bootstrap model selection) for 1 nearest neighbor
}
