% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_accs.R
\name{count_acc}
\alias{count_acc}
\title{Subroutine to compute average test accuracies ATA_k given correct-label ranks.}
\usage{
count_acc(ranks, K = length(counts), subKs = 1:length(counts))
}
\arguments{
\item{K}{Total number of labels in test set.}

\item{subKs}{The number of classes k for which we want to calculate ATA_k.}

\item{counts}{The ranks of the correct label for each tets instance.}
}
\description{
Subroutine to compute average test accuracies ATA_k given correct-label ranks.
}
