% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{d_piK}
\alias{d_piK}
\title{Function which computes derivative of piK}
\usage{
d_piK(mus, K, mc.reps = 10000)
}
\arguments{
\item{mus}{Mean mu, can be vectorized.}

\item{K}{Number of independent gaussians.}

\item{mc.reps}{Numerical resolution (larger for more precision.)}
}
\value{
A real
}
\description{
Function which computes derivative of piK
}
\examples{
d_piK(3, 10)
}

