% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do.call2}
\alias{do.call2}
\title{Similar to do.call, but allows extra arguments}
\usage{
do.call2(f, ll, ...)
}
\arguments{
\item{f}{function to be evaluated}

\item{ll}{list containing arguments}
}
\description{
Similar to do.call, but allows extra arguments
}
\examples{
do.call2(cbind, list(a = rep(0, 3), b = rep(1, 3), c = rep(2, 3)), a = c(1, 2, 3), d = c(4, 5, 4))
}

