% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenprism.R
\name{eigenprism_w}
\alias{eigenprism_w}
\title{Determines eigenprism weights}
\usage{
eigenprism_w(l, tol = 1e-04, sigma2 = FALSE)
}
\arguments{
\item{l}{Eigenvalues of X'X}

\item{tol}{Tolerance}

\item{sigma2}{Compute weights for estimating sigma2 (TRUE) or signal strength (FALSE)?}
}
\description{
Determines eigenprism weights
}
\references{
Janson 2015
}

