% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenprism.R
\name{eigenprisms}
\alias{eigenprisms}
\title{Confidence intervals for ||beta||^2 or sigma2 (see Janson 2015)}
\usage{
eigenprisms(X, Y, alpha = 0.05, sigma2 = FALSE)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{matrix response}

\item{alpha}{1-coverage of confidence interval}

\item{sigma2}{Compute weights for estimating sigma2 (TRUE) or signal strength (FALSE)?}
}
\description{
Confidence intervals for ||beta||^2 or sigma2 (see Janson 2015)
}
\references{
Janson 2015
}

