% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastRank}
\alias{fastRank}
\title{Computing the rank of the correct label given classification scores.}
\usage{
fastRank(pmat, Zs)
}
\arguments{
\item{pmat}{Matrix of margins m_y(x) for test instances, rows are observations and columns are labels.}

\item{Zs}{True labels.}
}
