% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastRank1nn}
\alias{fastRank1nn}
\title{Computing the rank of the Euclidean distance between observation and correct centroid vs all centroids.}
\usage{
fastRank1nn(Xm, Ym, Zs)
}
\arguments{
\item{Xm}{Matrix of centroids, rows are centroids and columns are dimensions.}

\item{Ym}{Data matrix, rows are observations and columns are dimensions.}

\item{Zs}{Assignments of observations to centroids.}
}
