% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_eigenprism}
\alias{filter_eigenprism}
\title{Filters out columns with estimated 0 signal size}
\usage{
filter_eigenprism(X, Y, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}
}
\value{
\code{filt}, logical vector
}
\description{
A \code{filter_method}.
}

