% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{fit_Bayes}
\alias{fit_Bayes}
\title{Optimal Bayes estimate given true Sigma_b}
\usage{
fit_Bayes(X, Y, X_te, Sigma_e, Sigma_t, Sigma_b, filt, mc.cores = 0, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{X_te}{Test class covariates}

\item{Sigma_e}{Response covariance}

\item{Sigma_t}{Autocorrelation over time}

\item{Sigma_b}{Prior covariance of each row of B (rows assumed independent).
Constrained to be diagonal!}

\item{filt}{Logical vector: which responses to use}
}
\value{
Coef matrix B
}
\description{
A \code{forward_method}.
NOTE: When used in pipeline, make sure to include
\code{forward_params = list(Sigma_b = Sigma_b, Sigma_e = Sigma_e, Sigma_t = Sigma_t)}
}

