% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{fit_elnet_CV}
\alias{fit_elnet_CV}
\title{Multivariate elastic net regression}
\usage{
fit_elnet_CV(X, Y, filt = rep(TRUE, dim(Y)[2]), alpha = 0,
  rule = "lambda.1se", mc.cores = 1, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{filt}{Logical vector: which responses to use}

\item{alpha}{Ridge (0) or Lasso (1)}

\item{rule}{Use \code{lambda.1se} or \code{lambda.min}}

\item{mc.cores}{number of mc cores}
}
\value{
B coefficient matrix
}
\description{
A \code{forward_method}.
Fits elastic net ridge separately to each column
}

