% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_basis.R
\name{gaussian_basis}
\alias{gaussian_basis}
\title{Formula for Gaussian-kernel based extrapolation basis functions}
\usage{
gaussian_basis(mu, sigma2, ks, n.sample = 10000)
}
\arguments{
\item{mu}{the mean of the basis function.}

\item{sigma2}{the variance parameter of the basis function.}

\item{ks}{the number of classes for which we want to compute the basis function.}

\item{n.sample}{number of sample points for computing Gaussian integrals}
}
\description{
Formula for Gaussian-kernel based extrapolation basis functions
}
