% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_method.R
\name{gaussian_kernel_cdf}
\alias{gaussian_kernel_cdf}
\title{Subroutine to get the KDE estimate of Pr[X > x] using Gaussian Kernels}
\usage{
gaussian_kernel_cdf(xs, x, bw = "bcv")
}
\arguments{
\item{xs}{Observations X_i}

\item{x}{Query point x}

\item{bw}{Bandwidth estimation method: see help("bw.nrd")}
}
\description{
Subroutine to get the KDE estimate of Pr[X > x] using Gaussian Kernels
}
