% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{gen_data}
\alias{gen_data}
\title{Generates data for a Bayesian multivariate regression model}
\usage{
gen_data(X, X_te, B, Sigma_b = NULL, Sigma_e, Sigma_t, n_te, ...)
}
\arguments{
\item{X}{Design matrix for training set}

\item{X_te}{Design matrix for test set}

\item{B}{coefficient matrix}

\item{Sigma_b}{Prior covariance of B}

\item{Sigma_e}{Prior within-row covariance of E}

\item{Sigma_t}{Prior within-column covariance of E}

\item{n_te}{Number of rows X_te to subsample to produce training set}
}
\value{
List with elements \code{X, Y, X_te, i_chosen, y_star, B,
Sigma_b, Sigma_e, Sigma_t}.  \code{y_star} are test set responses
while \code{i_chosen} are the indices corresponding rows of \code{X_te}.
}
\description{
See \code{\link{gen_params}} to generate X, B and covariance matrices
}

