% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{gen_params}
\alias{gen_params}
\title{Generates parameters for a Bayesian multivariate regression model}
\usage{
gen_params(n = 30, pY = 60, pX = 70, W_X = 2, s_e = 10, s_b = 1,
  df_b = 2, W_e = 2, rho_t = 0, L = 100, n_te = 10, ...)
}
\arguments{
\item{n}{Training set size}

\item{pY}{Dimension of response}

\item{pX}{Dimension of covariate}

\item{W_X}{Degrees of freedom multiplier for \code{Sigma_X}}

\item{s_e}{Noise standard error multipler}

\item{s_b}{Signal size multiplier}

\item{df_b}{Degrees of freedom for elements of \code{Sigma_b}}

\item{W_e}{Degrees of freedom multiplier for \code{Sigma_e}}

\item{rho_t}{Autocorrelation parameter for \code{Sigma_t}}

\item{L}{Number of test set candidates}

\item{n_te}{Test set size}
}
\value{
A list with elements \code{X, X_te, B, Sigma_b, Sigma_e,
Sigma_t, n_te}
}
\description{
Generates covariance matrices for X, E, and B,
then generates X, B, and test set X.
See \code{\link{gen_data}} to generate response Y.
\itemize{
  \item Covariance matrices are appropriated scaled Wishart matrices.
  \item \code{Sigma_X} generated with \code{W_X * pX} degrees of freedom.
  \item \code{Sigma_e} generated with \code{W_e * pY} degrees of freedom.
  \item \code{Sigma_t} generated as autoregressive covariance with parameter \code{rho_t}.
  \item \code{Sigma_b} diagonal with scaled (mean 1) chi-squared elements.
  \item Rows of \code{X} are normal \code{Sigma_X}.
  \item Noise matrix \code{E} has covariance \code{Sigma_t \%*\% Sigma_e}.
}
}

