% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_basis.R
\name{get_gaussian_basis_mat}
\alias{get_gaussian_basis_mat}
\title{Get a Gaussian-kernel based basis matrix for ClassExReg.}
\usage{
get_gaussian_basis_mat(max.mu, kernel_sigma, Ktrain, Ktarg, n.sample = 10000)
}
\arguments{
\item{max.mu}{maximum mean to use in the basis}

\item{kernel_sigma}{the sd. dev parameter of the basis function.}

\item{Ktrain}{the number of classes k for which we have computed average test accuracies ATA_k.}

\item{Ktarg}{The number of classes k for which we want to extrpolate the accuracy AGA_k (can be a vector).}

\item{n.sample}{number of sample points for computing Gaussian integrals}
}
\description{
Get a Gaussian-kernel based basis matrix for ClassExReg.
}
