% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{Ihat_LI}
\alias{Ihat_LI}
\title{Low-information estimator of Mutual Information}
\usage{
Ihat_LI(p, K, upper = 10, res = 1000)
}
\arguments{
\item{p}{Misclassification probability.}

\item{K}{Number of classes.}

\item{upper}{A guess on the upper bound of the mutual information (should be exponential in K.)}

\item{res}{Numerical resolution (larger for more precision.)}
}
\value{
The estimated mutual information.
}
\description{
Given a misclassification rate and number of classes, estimate the mutual information between X and Y.
}
\examples{
Ihat_LI(0.5, 5, 100)
}

