% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{inv_piK}
\alias{inv_piK}
\title{Function which inverts piK}
\usage{
inv_piK(p, K, upper = 10, res = 1000)
}
\arguments{
\item{p}{Probability}

\item{K}{Number of independent gaussians.}

\item{upper}{A guess on the upper bound of mu.}

\item{res}{Numerical resolution (larger for more precision.)}
}
\value{
The parameter mu such that \code{piK(mu, K) = p}.
}
\description{
\code{piK} computes the probability that a normal with mean mu is smaller than the max of K-1 other iid Gaussian.
This the the inverse function.
}
\examples{
p <- piK(3, 10)
inv_piK(p, 10, 5)
}

