% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAlg.R
\name{kron_v}
\alias{kron_v}
\title{Computes \code{(A \%x\% B) \%*\% v}}
\usage{
kron_v(A, B, cc)
}
\arguments{
\item{A}{matrix of size a x b}

\item{B}{matrix of size c x d}

\item{cc}{vector of length b * d}
}
\value{
vector of size a * c
}
\description{
Computes \code{(A \%x\% B) \%*\% v}
}
\examples{
a1 <- 5; a2 <- 3; b1 <- 3; b2 <- 4
A <- pracma::randn(a1, a2)
B <- pracma::randn(b1, b2)
cc <- rnorm(a2 * b2)
ans <- kron_v(A, B, cc)
ansn <- (A \%x\% B) \%*\% cc
f2(ans, ansn)
}

