% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lclapply}
\alias{lclapply}
\title{Calls mclapply and combines the end result using listcomb}
\usage{
lclapply(x, f, mc.cores = 0)
}
\arguments{
\item{x}{List of arguments}

\item{f}{Single-argument function}

\item{mc.cores}{Number of cores: 0 means use lapply}
}
\description{
Calls mclapply and combines the end result using listcomb
}
\examples{
lclapply(1:10, function(i) list(a = i, b = i^2), mc.cores = 1)
}

