% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listcomb}
\alias{listcomb}
\title{Converts list(list(fields)) into list(fields(list))}
\usage{
listcomb(dots)
}
\arguments{
\item{dots}{list of lists}
}
\description{
Converts list(list(fields)) into list(fields(list))
}
\examples{
listcomb(list(list(a = 5, b = 3), list(a = 2, b = 4)))
}

