% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mclapply0}
\alias{mclapply0}
\title{Either uses lapply or mclapply}
\usage{
mclapply0(x, f, mc.cores = 0)
}
\arguments{
\item{x}{List of arguments}

\item{f}{Single-argument function}

\item{mc.cores}{Number of cores: 0 means use lapply}
}
\description{
Either uses lapply or mclapply
}
\examples{
lclapply(1:10, function(i) list(a = i, b = i^2), mc.cores = 1)
}

