% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{meanexp}
\alias{meanexp}
\title{Exponentiate then take the mean}
\usage{
meanexp(v)
}
\arguments{
\item{v}{Values v_i}
}
\value{
Mean of exp(v_i)
}
\description{
Exponentiate then take the mean
}
\examples{
meanexp(c(-1, 3, 4))
mean(exp(c(-1, 3, 4)))
}

