% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mvrnorm2}
\alias{mvrnorm2}
\title{Generates normal matrix with both row and column correlations}
\usage{
mvrnorm2(Sigma_e, Sigma_t)
}
\arguments{
\item{Sigma_e}{(p x p) matrix, covariances between columns (or within a row)}

\item{Sigma_t}{(T x T) matrix, covariances between rows (or within a column)}
}
\value{
A (T x p) matrix
}
\description{
Multivariate normal with covariance \code{Sigma_e \%x\% Sigma_t}
}

