% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{obs_data_filter}
\alias{obs_data_filter}
\title{Extracts observed data from a list generated by gen_data}
\usage{
obs_data_filter(X, X_te, Y, y_star, ...)
}
\description{
Use with \code{do.call}
}
\examples{
pars <- gen_params()
dat <- do.call(gen_data, pars)
obs <- do.call(obs_data_filter, dat)
}

