% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAlg.R
\name{paramultiply}
\alias{paramultiply}
\title{Multiple core matrix multiplication}
\usage{
paramultiply(A, B, mc.cores = 3)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix with many columns}
}
\value{
The product
}
\description{
Given A and B, computes AB by splitting columns of B
}
\examples{
A <- randn(100, 100)
B <- randn(100, 1000)
C <- paramultiply(A, B, mc.cores = 3)
}

