% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{post_likes}
\alias{post_likes}
\title{Computing posterior likelihoods from predictive moments}
\usage{
post_likes(X_te, y_star, pre_moments, filt = rep(TRUE, dim(y_star)[2]),
  mc.cores = 0, ...)
}
\arguments{
\item{X_te}{Test class covariates}

\item{y_star}{Responses to be classified}

\item{pre_moments}{Predictive means and covariances for test classes}

\item{filt}{Logical vector: which responses to use}

\item{mc.cores}{Number of parallel cores}
}
\value{
Matrix of log-likelihoods, rows = responses, columns = classes
}
\description{
Computes gaussian likelihoods for every (response, class) combo.
}

