% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{post_predictive}
\alias{post_predictive}
\title{Posterior predictive means}
\usage{
post_predictive(X, Y, X_te, Sigma_e, Sigma_b, Sigma_t = eye(dim(X)[1]),
  naive = FALSE, mc.cores = 0, ...)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{X_te}{covariates for responses to be predicted}

\item{Sigma_e}{Covariance within given time (row) of Y}

\item{Sigma_b}{Prior covariance of a column of B}

\item{Sigma_t}{Autocorrelation of a single response (column) of Y over time}

\item{naive}{Do the naive computation (for testing)}

\item{mc.cores}{How many cores to use in parallel (0 for single-core)}
}
\value{
a list of list, one sublist per row of X_te,
containing posterior mean and covariance
}
\description{
Returns a list with entry for every row of Xte, which is a list with mu, cov
}
\examples{
n <- 100; pX <- 20; pY <- 30; L <- 10
X <- randn(n, pX)
X_te <- randn(L, pX)
B <- randn(pX, pY)
Sigma_b <- eye(pY)
Sigma_e <- cor(randn(3 * pY, pY))
Sigma_t <- cor(randn(3 * n, n))
E <- sqrtm(Sigma_t) \%*\% randn(n, pY) \%*\% sqrtm(Sigma_e)
Y <- X \%*\% B + E
E2 <- randn(L, pY) \%*\% sqrtm(Sigma_e)
Y_te <- X_te \%*\% B + E2
res <- post_predictive(X, Y, X_te, Sigma_e, Sigma_b, Sigma_t)
Bols <- solve(t(X) \%*\% X, t(X) \%*\% Y)
Yh <- do.call(rbind, listcomb(res)$Mu)
f2(X_te \%*\% Bols, Y_te)
f2(Yh, Y_te)
}

