% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{pre_Bayes}
\alias{pre_Bayes}
\title{Predictive distribution given true or estimated Sigma_b}
\usage{
pre_Bayes(X, Y, X_te, Sigma_e, Sigma_t, Sigma_b, filt, mc.cores = 0, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{X_te}{Test class covariates}

\item{Sigma_e}{Response covariance}

\item{Sigma_t}{Autocorrelation over time}

\item{Sigma_b}{Prior covariance of each row of B (rows assumed independent).
Constrained to be diagonal!}

\item{filt}{Logical vector: which responses to use}
}
\value{
A list with moments
}
\description{
A \code{backward_method}.
Computes a different covariance for each response due to posterior variance of B.
NOTE: To obtain Bayes rule (oracle) using pipeline, include
\code{backward_params = list(Sigma_b = Sigma_b, Sigma_e = Sigma_e, Sigma_t = Sigma_t)}
}
\examples{
pars <- gen_params(n = 50, pY = 20, pX = 30, n_te = 100)
dat <- do.call(gen_data, pars)
zattach(dat)
"Empirical Bayes"
ep_res <- identification_pipeline1(X, Y, X_te, y_star, i_chosen,
  filter_method = filter_eigenprism, forward_method = fit_elnet_CV,
  Sigma_e_method = residual_offdiag, Sigma_t_method = assume_iid,
  Sigma_b_method = use_eigenprism,
  backward_method = pre_Bayes,
  scoring_method = topk_score,
  mc.cores = 3)
ep_res$score
"Oracle Bayes rule"
bayesrule <- identification_pipeline1(X, Y, X_te, y_star, i_chosen,
  filter_method = no_filter, forward_method = fit_elnet_CV,
  Sigma_e_method = residual_offdiag, Sigma_t_method = assume_iid,
  backward_method = pre_Bayes,
  backward_params = list(Sigma_b = Sigma_b, Sigma_e = Sigma_e,
    Sigma_t = Sigma_t),
  scoring_method = topk_score,
  mc.cores = 3)
bayesrule$score
}

