% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_method.R
\name{raccs}
\alias{raccs}
\title{Subroutine to get all of the KDE estimates of the favorability for each test instance}
\usage{
raccs(pmat, i_chosen, bw)
}
\arguments{
\item{pmat}{Matrix of margins m_y(x) for test instances, rows are observations and columns are labels.}

\item{i_chosen}{The true labels for the observations.}

\item{bw}{Bandwidth estimation method: see help("bw.nrd")}
}
\description{
Subroutine to get all of the KDE estimates of the favorability for each test instance
}
