% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{resample_misclassification}
\alias{resample_misclassification}
\title{Probability of misclassification if we resampled m out of K classes}
\usage{
resample_misclassification(plikes, i_chosen, m = dim(plikes)[2],
  replace = FALSE, ...)
}
\arguments{
\item{plikes}{Matrix of posterior likelihoods, rows = responses, columns = classes}

\item{i_chosen}{True classes}

\item{m}{number of classes resampled}

\item{replace}{TRUE or FALSE: whether to sample with replacement}
}
\value{
Misclassification probability
}
\description{
A \code{scoring_method}.
}

