% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{residual_offdiag}
\alias{residual_offdiag}
\title{Covariance estimation using residuals}
\usage{
residual_offdiag(X, Y, B, filt = rep(TRUE, dim(Y)[2]), shrink = 0.5,
  mc.cores = 1, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{B}{fitted coefficient matrix}

\item{filt}{Logical vector: which responses to use}

\item{shrink}{shrinkage factor}

\item{mc.cores}{parallelization}
}
\value{
Estimate for covariance matrix
}
\description{
A \code{Sigma_e_method}.
Estimates within-row covariances (or \code{Sigma_e}) 
using sample covariance of residuals,
applying off-diagonal shrinkage.
}

