% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{samp_moments}
\alias{samp_moments}
\title{Sampling distribution of ridge regression}
\usage{
samp_moments(X, Y, Sigma_e, lambdas, Sigma_t = eye(dim(X)[1]),
  computeCov = TRUE, matrix = TRUE, ...)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{Sigma_e}{Covariance within given time (row) of Y}

\item{lambdas}{Ridge regularization per column of B}

\item{Sigma_t}{Autocorrelation of a single response (column) of Y over time}

\item{computeCov}{Return the posterior covariance of B?}

\item{matrix}{Return B as a matrix? (or vector)}
}
\value{
a matrix/vector (if \code{computeCov = FALSE})
or a list (if \code{computeCov = TRUE})
}
\description{
Sampling distribution of ridge regression
}

