% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{samp_predictive}
\alias{samp_predictive}
\title{Frequentist predictive distributions for ridge regression}
\usage{
samp_predictive(X, Y, X_te, Sigma_e, lambdas, Sigma_t = eye(dim(X)[1]),
  mc.cores = 0, ...)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{X_te}{covariates for responses to be predicted}

\item{Sigma_e}{Covariance within given time (row) of Y}

\item{lambdas}{Ridge regularization per column of B}

\item{Sigma_t}{Autocorrelation of a single response (column) of Y over time}

\item{mc.cores}{How many cores to use in parallel (0 for single-core)}
}
\value{
a list of list, one sublist per row of X_te,
containing posterior mean and covariance
}
\description{
Returns a list with entry for every row of Xte, which is a list with mu, cov
}

