% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAlg.R
\name{split_vec}
\alias{split_vec}
\title{Splits a vector into approximately equally-sized parts}
\usage{
split_vec(v, n.splits)
}
\arguments{
\item{v}{Vector to be split}

\item{n.splits}{Number of splits}
}
\value{
A list of vectors
}
\description{
Splits a vector into approximately equally-sized parts
}
\examples{
split_vec(1:100, 6)
}

