% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_accs.R
\name{sub_accuracies}
\alias{sub_accuracies}
\title{Compute average test accuracies ATA_k.}
\usage{
sub_accuracies(pmat, i_chosen, subKs = 1:ncol(pmat))
}
\arguments{
\item{pmat}{Matrix of margins m_y(x) for test instances, rows are observations and columns are labels.}

\item{i_chosen}{True labels.}

\item{subKs}{The number of classes k for which we want to calculate ATA_k.}
}
\description{
Compute average test accuracies ATA_k.
}
