% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_accs.R
\name{sub_accuracies_1nn}
\alias{sub_accuracies_1nn}
\title{Compute average test accuracies ATA_k for 1-nearest neighbor.}
\usage{
sub_accuracies_1nn(Xm, Ym, i_chosen, subKs = 1:nrow(Xm))
}
\arguments{
\item{Xm}{Matrix of centroids, rows are centroids and columns are dimensions.}

\item{Ym}{Data matrix, rows are observations and columns are dimensions.}

\item{i_chosen}{True labels.}

\item{subKs}{The number of classes k for which we want to calculate ATA_k.}
}
\description{
Compute average test accuracies ATA_k for 1-nearest neighbor.
}
