% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{sum_score}
\alias{sum_score}
\title{Sum of correct posterior probabilities}
\usage{
sum_score(plikes, i_chosen, ...)
}
\arguments{
\item{plikes}{Matrix of posterior likelihoods, rows = responses, columns = classes}

\item{i_chosen}{True classes}
}
\value{
Sum of correct probabilities
}
\description{
A \code{scoring_method}.
}

