% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAlg.R
\name{tkron_d_kron}
\alias{tkron_d_kron}
\title{Computes \code{t(A \%x\% B) \%*\% diag(d) \%*\% (A \%*\% B)}}
\usage{
tkron_d_kron(A, B, d)
}
\arguments{
\item{A}{matrix of size a x b}

\item{B}{matrix of size c x d}

\item{d}{vector: represents diagonal matrix of size a * c}
}
\value{
square matrix of size a * c
}
\description{
Computes \code{t(A \%x\% B) \%*\% diag(d) \%*\% (A \%*\% B)}
}
\examples{
a1 <- 5; a2 <- 3; b1 <- 4; b2 <- 6
A <- pracma::randn(a1, a2)
B <- pracma::randn(b1, b2)
cc <- rnorm(a1 * b1)
ans <- tkron_d_kron(A, B, cc)
ansn <- t(A \%x\% B) \%*\% diag(cc) \%*\% (A \%x\% B)
f2(ans, ansn)
}

