% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{topk_score}
\alias{topk_score}
\title{Counting correct classifications within top-K rankings}
\usage{
topk_score(plikes, i_chosen, k = 1, ...)
}
\arguments{
\item{plikes}{Matrix of posterior likelihoods, rows = responses, columns = classes}

\item{i_chosen}{True classes}

\item{k}{Number of top candidates for each response}
}
\value{
Number correct
}
\description{
A \code{scoring_method}.
}

