% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{topk_scores}
\alias{topk_scores}
\title{Counting correct classifications within top-K rankings}
\usage{
topk_scores(plikes, i_chosen, max_k = ncol(plikes), ...)
}
\arguments{
\item{plikes}{Matrix of posterior likelihoods, rows = responses, columns = classes}

\item{i_chosen}{True classes}

\item{max_k}{max number of top candidates for each response}
}
\value{
Number correct for K = 1...max_k
}
\description{
A \code{scoring_method}.
}

