% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{use_eigenprism}
\alias{use_eigenprism}
\title{Uses eigenprism to estimate prior covariances}
\usage{
use_eigenprism(X, Y, filt = rep(TRUE, dim(Y)[2]), ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{filt}{Logical vector: which responses to use}
}
\value{
Estimate for Sigma_b, a diagonal matrix
}
\description{
A \code{Sigma_b_method}.
It is recommended to use filter_eigenprism in conjunction with this method.
}

