% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{use_norms_of_Bhat}
\alias{use_norms_of_Bhat}
\title{Uses norms of a previously fitted B to set prior}
\usage{
use_norms_of_Bhat(B, ...)
}
\arguments{
\item{B}{fitted coefficient matrix}
}
\value{
Estimate for Sigma_b, a diagonal matrix
}
\description{
A \code{Sigma_b_method}.
}

