% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zattach}
\alias{zattach}
\title{Similar to attach, but overwrites global}
\usage{
zattach(ll)
}
\arguments{
\item{ll}{list with named values}
}
\description{
Similar to attach, but overwrites global
}
\examples{
zattach(list(a = 5))
a
}

